+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+## [2.1.0] - 2026-08-37
+
+- Improvement: Add functions to assist with nuitka
+- Improvement: Add functions to check if module(s) can be imported
+- Improvement: Add function to read pyproject file
+- Improvement: Add function to check if module is one file
+- Improvement: Add functions to overwrite one folder with another
+
 ## [2.0.1] - 2026-01-31
 
 - [Fix] Fissing `packaging` dependency marker in 2.0.0